+++ R/bernoulli_naive_bayes.R
@@ -1,31 +1,30 @@
 bernoulli_naive_bayes <- function (x, y, prior = NULL, laplace = 0, ...)  {
 
     if (!is.factor(y) & !is.character(y) & !is.logical(y))
-        stop("bernoulli_naive_bayes(): y has to be either a factor or character or logical vector", call. = FALSE)
-
+        stop("multinomial_naive_bayes(): y has to be either a factor or character or logical vector", call. = FALSE)
     if (!is.factor(y))
         y <- factor(y)
     levels <- levels(y)
     nlev <- nlevels(y)
     vars <- colnames(x)
-
-    if (nlev < 2)
-        warning("bernoulli_naive_bayes(): y has less than two classes. ", call. = FALSE)
-
-    if (is.null(vars)) {
-        xname <- deparse(substitute(x))
-        stop(paste0("bernoulli_naive_bayes(): Column names in the matrix x are required.\n",
-                    "       Consider paste0(\"V\", 1:ncol(", xname, ")) as column names \n",
-                    "       in both train and test datasets."), call. = FALSE)
-    }
-    if (class(x)[1] != "matrix") {
-        stop("bernoulli_naive_bayes(): x has to be a numeric 0-1 matrix. ", call. = FALSE)
+    class_x <- class(x)[1]
+    use_Matrix <- class_x %in% .matrix_classes
+    if (!is.matrix(x) & !use_Matrix) {
+        warning("bernoulli_naive_bayes(): x was coerced to matrix.", call. = FALSE)
         x <- as.matrix(x)
         if (mode(x) != "numeric")
-            stop("bernoulli_naive_bayes(): x has to contain numeric columns with 0-1 values. ",
-                 "Please consider coercing features to numeric 0-1 or using the general \"naive_bayes\"",
-                 "function, which models \"character\", \"factor\" or \"logical\" variables with two levels with Bernoulli.", call. = FALSE)
+            stop("bernoulli_naive_bayes(): x must be a matrix/dgCMatrix with with numeric {0,1} columns.", call. = FALSE)
     }
+    if (use_Matrix) {
+        if (!"Matrix" %in% rownames(utils::installed.packages()))
+            stop("bernoulli_naive_bayes(): please install \"Matrix\" package.")
+        if (class_x != "dgCMatrix")
+            stop("bernoulli_naive_bayes(): dgCMatrix class from the Matrix package is only supported.", call. = FALSE)
+    }
+    if (nlev < 2)
+        stop("bernoulli_naive_bayes(): y must contain at least two classes. ", call. = FALSE)
+    if (is.null(vars))
+        stop("bernoulli_naive_bayes(): x must have unique column names.\n", call. = FALSE)
     NAy <- anyNA(y)
     NAx <- anyNA(x)
     if (NAy) {
@@ -34,7 +33,7 @@ bernoulli_naive_bayes <- function (x, y, prior = NULL, laplace = 0, ...)  {
         warning(paste0("bernoulli_naive_bayes(): y contains ", len_na, " missing",
                        ifelse(len_na == 1, " value", " values"), ". ",
                        ifelse(len_na == 1, "It is", "They are"),
-                       " not included (together with the corresponding instances in x) ",
+                       " not included (also the corresponding rows in x) ",
                        "into the estimation process."), call. = FALSE)
         y <- y[!na_y_bool]
         x <- x[!na_y_bool, ]
@@ -50,7 +49,7 @@ bernoulli_naive_bayes <- function (x, y, prior = NULL, laplace = 0, ...)  {
     y_counts <- stats::setNames(tabulate(y), levels)
     y_min <- y_counts < 1
     if (any(y_min))
-        stop(paste0("bernoulli_naive_bayes(): y variable has to contain at least ",
+        stop(paste0("bernoulli_naive_bayes(): y variable must contain at least ",
                     "one observation per class for estimation process.",
                     " Class ", paste0(levels[y_min], collapse =  ", "),
                     " has less than 1 observation."), call. = FALSE)
@@ -58,19 +57,42 @@ bernoulli_naive_bayes <- function (x, y, prior = NULL, laplace = 0, ...)  {
         prior <- prop.table(y_counts)
     } else {
         if (length(prior) != nlev)
-            stop(paste0("bernoulli_naive_bayes(): Vector with prior probabilities should have ",
+            stop(paste0("bernoulli_naive_bayes(): vector with prior probabilities must have ",
                         nlev, " entries"))
         prior <- stats::setNames(prior / sum(prior), levels)
     }
     if (!NAx) {
-        prob1 <- t((rowsum(x, y, na.rm = TRUE) + laplace) /  (y_counts + laplace * 2))
-    } else {
-        n <- rowsum((!na_x_bool) * 1, y)
-        if (any(n < 2)) {
-            warning(paste0("bernoulli_naive_bayes(): x has to contain at least one ",
-                           "non-missing observation per class for estimation process."), call. = FALSE)
+        prob1 <- if (use_Matrix) {
+            params <- lapply(levels, function(lev) {
+                Matrix::colSums(x[y == lev, , drop = FALSE], na.rm = TRUE) + laplace })
+            params <- do.call("rbind", params)
+            t(params / (y_counts + laplace * 2))
+        } else {
+            t((rowsum.default(x, y, na.rm = TRUE) + laplace) /  (y_counts + laplace * 2))
         }
-        prob1 <- t((rowsum(x, y, na.rm = TRUE) + laplace) /  (n + laplace * 2))
+    } else {
+        n <- if (use_Matrix) {
+            n_per_feature <- lapply(levels, function(lev) {
+                not_na_x <- (!na_x_bool) * 1
+                Matrix::colSums(not_na_x[y == lev, , drop = FALSE], na.rm = TRUE)})
+            n_per_feature <- do.call("rbind", n_per_feature)
+            rownames(n_per_feature) <- levels
+            n_per_feature
+        } else {
+            rowsum.default((!na_x_bool) * 1, y)
+        }
+        if (any(n == 0)) {
+            warning(paste0("bernoulli_naive_bayes(): x should contain at least one ",
+                           "non-missing observation per class for parameter estimation purposes."), call. = FALSE)
+        }
+        prob1 <- if (use_Matrix) {
+            params <- lapply(levels, function(lev) {
+                Matrix::colSums(x[y == lev, , drop = FALSE], na.rm = TRUE) + laplace })
+            params <- do.call("rbind", params)
+            t(params / (n + laplace * 2))
+        } else {
+            t((rowsum.default(x, y, na.rm = TRUE) + laplace) /  (n + laplace * 2))
+        }
     }
     if (any(prob1 == 0)) {
         nempty <- length(which(prob1 == 0, arr.ind = TRUE)[ ,1])
@@ -88,10 +110,14 @@ predict.bernoulli_naive_bayes <- function(object, newdata = NULL, type = c("clas
 
     if (is.null(newdata))
         newdata <- object$data$x
-    if (!is.matrix(newdata))
-        stop("predict.bernoulli_naive_bayes(): \"newdata\" has to be a matrix with at least one row and two columns.", call. = FALSE)
-    if (mode(newdata) != "numeric")
-        stop("predict.bernoulli_naive_bayes(): \"newdata\" has to be a matrix with numeric 0-1 values.", call. = FALSE)
+    class_x <- class(newdata)[1]
+    use_Matrix <- class_x == "dgCMatrix"
+    if (!is.matrix(newdata) & !use_Matrix)
+        stop("predict.multinomial_naive_bayes(): newdata has to be numeric matrix or dgCMatrix (Matrix package) with at least one row and two named columns.", call. = FALSE)
+    if (is.matrix(newdata) & mode(newdata) != "numeric")
+        stop("predict.multinomial_naive_bayes(): newdata has to be a numeric matrix.", call. = FALSE)
+    if (use_Matrix & !"Matrix" %in% rownames(utils::installed.packages()))
+        stop("predict.multinomial_naive_bayes(): please install Matrix package", call. = FALSE)
 
     type <- match.arg(type)
     lev <- object$levels
@@ -103,37 +129,24 @@ predict.bernoulli_naive_bayes <- function(object, newdata = NULL, type = c("clas
     n_tables <- ncol(prob1)
     prob1 <- prob1[ ,features, drop = FALSE]
     n_features <- length(features)
+    n_features_newdata <- ncol(newdata)
 
     if (n_features == 0) {
+        warning(paste0("predict.multinomial_naive_bayes(): no feature in newdata corresponds to ",
+                       "features defined in the object. Classification is based on prior probabilities"), call. = FALSE)
         if (type == "class") {
-            warning(paste0("predict.bernoulli_naive_bayes(): ",
-                           "No feature in the newdata corresponds to ",
-                           "probability tables in the object. ",
-                           "Classification is done based on the prior probabilities"),
-                    call. = FALSE)
-            return(factor(rep(lev[which.max(prior)], n_obs),
-                          levels = lev))
+            return(factor(rep(lev[which.max(prior)], n_obs), levels = lev))
         } else {
-            warning(paste0("predict.bernoulli_naive_bayes(): ",
-                           "No feature in the newdata corresponds to ",
-                           "probability tables in the object. ",
-                           "Posterior probabilities are equal to prior probabilities."),
-                    call. = FALSE)
-            return(matrix(prior, ncol = n_lev, nrow = n_obs,
-                          byrow = TRUE, dimnames = list(NULL, lev)))
+            return(matrix(prior, ncol = n_lev, nrow = n_obs, byrow = TRUE, dimnames = list(NULL, lev)))
         }
     }
     if (n_features < n_tables) {
-        warning(paste0("predict.bernoulli_naive_bayes(): Only ", n_features, " feature(s) out of ", n_tables,
-                       " defined in the naive_bayes object \"", substitute(object),
-                       "\" are used for prediction\n"), call. = FALSE)
+        warning(paste0("predict.multinomial_naive_bayes(): only ", n_features, " feature(s) in newdata could be matched ",
+                       "with ", n_tables, " feature(s) defined in the object."), call. = FALSE)
     }
-    if (ncol(newdata) > n_features) {
-        warning(paste0("predict.bernoulli_naive_bayes(): ",
-                       "More features in the newdata are provided ",
-                       "as there are probability tables in the object. ",
-                       "Calculation is performed based on features to be found in the tables."),
-                call. = FALSE)
+    if (n_features_newdata > n_features) {
+        warning(paste0("predict.multinomial_naive_bayes(): newdata contains feature(s) that could not be matched ",
+                       "with (", n_features, ") feature(s) defined in the object. Only matching features are used for calculation."), call. = FALSE)
         newdata <- newdata[ ,features, drop = FALSE]
     }
     if (object$laplace == 0) {
@@ -146,19 +159,21 @@ predict.bernoulli_naive_bayes <- function(object, newdata = NULL, type = c("clas
     lprob0 <- log(1 - prob1)
     NAs <- anyNA(newdata)
     if (NAs) {
-        ind_na <- which(is.na(newdata), arr.ind = TRUE)
+        ind_na <- if (use_Matrix) Matrix::which(is.na(newdata), arr.ind = TRUE) else which(is.na(newdata), arr.ind = TRUE)
         len_na <- nrow(ind_na)
-        warning(paste0("predict.bernoulli_naive_bayes(): ", len_na, " missing",
-                       ifelse(len_na == 1, " value", " values"), " discovered in the newdata. ",
-                       ifelse(len_na == 1, "It is", "They are"),
-                       " not included into the calculation."), call. = FALSE)
+        warning(paste0("predict.bernoulli_naive_bayes(): ", len_na, " missing", ifelse(len_na == 1, " value", " values"),
+                       " discovered in the newdata. ", ifelse(len_na == 1, "It is", "They are"), " not included into the calculation."), call. = FALSE)
         ind_obs <- ind_na[ ,1]
         ind_var <- ind_na[ ,2]
         newdata[ind_na] <- 1
         neutral <- do.call(rbind, tapply(ind_var, ind_obs, function(x) rowSums(lprob1[ ,x, drop = FALSE])))
         ind_obs <- sort(unique(ind_obs))
     }
-    post <- tcrossprod(newdata, lprob1) + tcrossprod(1 - newdata, lprob0)
+    if (use_Matrix) {
+        post <- Matrix::tcrossprod(newdata, lprob1) + Matrix::tcrossprod(1 - newdata, lprob0)
+    } else {
+        post <- tcrossprod(newdata, lprob1) + tcrossprod(1 - newdata, lprob0)
+    }
     if (NAs) {
         post[ind_obs, ] <- post[ind_obs, ] - neutral
     }
@@ -179,7 +194,7 @@ predict.bernoulli_naive_bayes <- function(object, newdata = NULL, type = c("clas
             return(post)
         }
         else {
-            return(apply(post, 2, function(x) { 1 / rowSums(exp(post - x)) }))
+            return(apply(post, 2, function(x) { 1 / if (use_Matrix) Matrix::rowSums(exp(post - x)) else rowSums(exp(post - x)) }))
         }
     }
 }